requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
PyPDF2
nltk
chromadb
python-dotenv
python-multipart
anthropic